Fix categorical loss calculation and update model references in turbine_limited.py

+++ hephaestus/single_row_models/training.py
@@ -173,7 +173,7 @@ class MaskedTabularModeling(L.LightningModule):
         )
 
         categorical_loss = self.categorical_loss_fn(
-            predicted_categorical, actual_categorical
+            predicted_categorical, actual_categorical.long()
         )
         return numeric_loss + categorical_loss
